Lab_prometheus_snmp_exporter/requirements.txt
prometheus-client
python3-netsnmp